+++ Plugins/PostgreSQLDriverPlugin/PostgreSQLPluginDriver.swift
@@ -85,7 +85,7 @@ class PostgreSQLPluginDriver: LibPQBackedDriver, @unchecked Sendable {
         }
     }
 
-    private func includesMaterializedViews() -> Bool {
+    func includesMaterializedViews() -> Bool {
         catalogPresence?.hasMaterializedViews ?? versionedCapabilities.hasMaterializedViewsCatalog
     }